refactor: prepare patientVaccination to activity switch
