+++ src/routes.tsx
@@ -43,7 +43,7 @@ const Routes = () => (
             <Route path="/PatientDatabase/PatientLabTest" component={PatientLabTest}/>
             <Route path="/PatientDatabase/PatientTherapy" component={PatientActivityContainer}/>
             <Route path="/PatientDatabase/PatientExamination" component={PatientActivityContainer}/>
-            <Route path="/PatientDatabase/PatientVaccine" component={PatientVaccine}/>
+            <Route path="/PatientDatabase/PatientVaccine" component={PatientActivityContainer}/>
             <Route path="/PatientDatabase/Vaccine" component={Vaccine}/>
             <Route path="/PatientDatabase/opd" component={PatientActivityContainer}/>
             <Route path="/PatientDatabase/new-opd" component={PatientActivityContainer}/>
